Changed mssql/query.js formatResults function
In order to return the result we disabled the BULKUPDATE and  allowed to run as a single update Query. On format result if isUpdateQuery we return the first row witch is the data as it was updated on the server. If its insert Query and it has a column that is TIMESTAMP then we return that column so we can have rowversioning